ai-assistant-mvp/backend/pyproject.toml
[tool.poetry]
name = "ai-assistant-mvp"
version = "0.1.0"
description = "AI Assistant MVP backend"
authors = ["You <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {version = "^0.30.0", extras = ["standard"]}
python-dotenv = "^1.0.1"
passlib = {version = "^1.7.4", extras=["bcrypt"]}
pyjwt = "^2.9.0"
sqlite-utils = "^3.36"
sqlmodel = "^0.0.21"
httpx = "^0.27.0"
aiosmtplib = "^3.0.1"
google-auth = "^2.27.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
